+++ projects/engine_opengl/nex/opengl/mesh/IndexBuffer.cpp
@@ -5,7 +5,7 @@
 namespace nex
 {
 
-	IndexBuffer::IndexBuffer(const void* data, unsigned int count, IndexElementType type) : IndexBuffer()
+	IndexBuffer::IndexBuffer(const void* data, size_t count, IndexElementType type) : IndexBuffer()
 	{
 
 		fill(data, count, type);
